Switching between dialogue and fights work now
