+++ text adventure project/dialogue/dg_test_001.json
@@ -1,0 +1,28 @@
+{
+    "text": "You walk up to the rocks",
+    "options": {
+        "eat a rock": {
+            "text": "Its crunchy",
+            "options": {
+                "eat more rocks": {
+                    "text": "you keep shoveling more and more rocks into your gaping maw. You are a rock friend. You Win!"
+                },
+                "become god": {
+                    "text": "you need to eat more rocks for that.[warp 1]"
+                }
+            }
+        },
+        "leave": {
+            "text": "you leave and find yourself at the bridge",
+            "options": {
+                "walk across": {
+                    "text": "you slip on a rock and break your neck. You should have eaten more rocks :("
+                },
+                "turn back and eat more rocks": {
+                    "text": "please stop"
+                }
+            }
+        }
+    },
+    "msgid": 1
+}
