--- text adventure project/dialogue/dialogue_001.json
@@ -1,35 +1,0 @@
-{
-    "text": "You walk up to the rocks",
-    "options": {
-        "eat a rock": {
-            "text": "Its crunchy",
-            "options": {
-                "eat more rocks": {
-                    "text": "you keep shoveling more and more rocks into your gaping maw. You are a rock friend. You Win!",
-                    "found": false
-                },
-                "become god": {
-                    "text": "you need to eat more rocks for that.[warp 1]",
-                    "found": false
-                }
-            },
-            "found": false
-        },
-        "leave": {
-            "text": "you leave and find yourself at the bridge",
-            "options": {
-                "walk across": {
-                    "text": "you slip on a rock and break your neck. You should have eaten more rocks :(",
-                    "found": false
-                },
-                "turn back and eat more rocks": {
-                    "text": "please stop",
-                    "found": false
-                }
-            },
-            "found": false
-        }
-    },
-    "found": false,
-    "msgid": 1
-}
