--- text adventure project/dialogue/dialogue_002.json
@@ -1,73 +1,0 @@
-{
-    "text": "Mr. Tran: Okay, I’m ready now. What do you need?",
-    "options": {
-        "I need a schedule": {
-            "text": "Mr Tran rubs the sleep off his eyes and looks at you, perplexed.\n\nMr Tran: Huh? Why don't you have one? Registration was two days ago. You should have one already.\n\nYou tell him that you went to registration, and all they did was shrug and tell you to go home.\n\nMr Tran: Hmmm. Don't worry, I can solve that no problem.\n\nHe pulls open a drawer in his desk and takes out a big red crayon. He then flips over some paper already on his desk and scribbles a schedule onto it.\n\nMr Tran: Let me just pull up your schedule on Aeries here.... one moment.... there! Here you go!\n\nAfter quickly giving it to you, he begins to lean back and return to his slumber.[unlock 2][lock 3][warp 1]",
-            "found": false,
-            "msgid": 3
-        },
-        "(read the schedule)": {
-            "text": "You take a look at the schedule. According to this, your first period is \"learning time\", located in \"classroom.\" You have a feeling this might not be the most credible schedule.[lock 2]",
-            "found": false,
-            "msgid": 2,
-            "locked": true,
-            "options": {
-                "(look at his screen)": {
-                    "text": "You peer over the desk and take a look at his computer screen. It's a youtube video, 3 hours into a 10 hour compilation of satisfying soap cutting. It's the only tab open.[lock 4][warp 2]",
-                    "found": false,
-                    "msgid": 4
-                },
-                "(tap mr tran to wake him up)": {
-                    "text": "You give Mr Tran a tap on the shoulder. His eyes slowly shift open, and he leans back up in his chair. You feel like you hear the noise of scraping concrete as he does so.\n\nMr Tran: What. Do. You. WANT FROM ME? Huh? Huh? I gave you your schedule and everything, stop bothering me![warp 5]"
-                },
-                "(kick the desk loudly)": {
-                    "text": "You kick the desk, making a loud thud. Mr Tran jumps back up from his seat and gasps loudly. His pale expression of horror quickly turns to a red one of rage as he looks to you.\n\nMr Tran: WHAT!? WHAT COULD YOU POSSBLY WANT!? I gave you your schedule![warp 5]"
-                }
-            }
-        },
-        "Were you asleep just now?": {
-            "text": "Mr Tran: I think you're mistaken. I was merely practicing my circadian meditation. It's important for maintaining my iron will required to be a counselor. You wouldn't understand.\n\nDespite his words, he looks at you half squinting with huge purple eye bags under his eyes. He reaches for a white coffee mug labeled with the text \"the most counselor ever\" written on it. He looks inside and makes a slight frown at the sight of it being empty.[warp 1]",
-            "found": false
-        },
-        "You suck, eat my pant.": {
-            "text": "Mr Tran: What did I do to deserve this.[warp 1]",
-            "found": false
-        },
-        "(leave)": {
-            "text": "",
-            "found": false
-        },
-        ">fake schedule<": {
-            "text": "",
-            "options": {
-                "This schedule isn't real Mr Tran": {
-                    "text": "Mr Tran: Uuugggghhhhhh. Fiiiiinne, I'll give you a better schedule since your so picky. But, at least do your do-diligence and get me a cup of coffee. I'm too tired too properly get any work done. Especially at this ungodly hour.\n\nYou look at the clock. It's 10:05 AM. It can't be that ungodly.",
-                    "options": {
-                        "How hard is it to just get my schedule?": {
-                            "text": "Mr Tran: Well, you see, that's all up to perspective, now isn't it? Apparently it's very hard for YOU to get your schedule, hence why you're coming to me. And what I'm telling you is that you need to get me a coffee. Got it?"
-                        },
-                        "You know what, fine. I'll get you a coffee.": {
-                            "text": "Mr Tran: That's what I like to hear. I like my coffee with equal parts "
-                        },
-                        "I'm not getting you a coffee.": {
-                            "text": ""
-                        },
-                        "Sure thing! How do you like your coffee?": {
-                            "text": ""
-                        },
-                        "Can't you get the coffee yourself?": {
-                            "text": ""
-                        },
-                        "How about instead you coff-up my schedule BUD?": {
-                            "text": ""
-                        }
-                    }
-                }
-            },
-            "locked": true,
-            "msgid": 5
-        }
-    },
-    "found": false,
-    "msgid": 1
-}
