+++ text adventure project/overview.txt
@@ -23,6 +23,20 @@ mr marestaing vr
 
 ##############
 
-You arrive at PHS without a schedule, so you go to Mr. Tran to find out what it is. 
-You go to his office, he'll only give you the schedule if you get him a coffee.
-The only area you can go is to the lunch area because geese block the way elsewhere.
+- Mr Trans office, tells you to get coffee if you want your schedule
+    - Has painkillers in the trash
+- When leaving the Admin building, first confrontation with Mr Chen.
+- Enter lunch area
+    - Cafeteria rat problem. They have a coffee machine but no ingredients.
+        - Will give you slop if you deal with the rats (rats schmats)
+    - Garden has Michealsen in the dirt.
+        - Some animal (he thinks its bunnies) is eating his crops
+        - Using the camo suit and hiding will make a little rat with a tiny bag come out and steal the crops. Following it leads to the rat den.
+    - Military recruitment table
+        - Sells stuff
+        - Sells the camo shirt, which is a necessary item to progress
+    - Guarding both the north and south paths to the rest of the school are blocked by the security guard and geese respectively.
+    - Biggy D statue
+        - Offering a coin gives you something
+    - Disgusting trash can
+        - Has an item in it
